refactored data viz components
